v.dissolve: Test for int/str versus float/double using a lib function
Assuming that the intention was to disallow floats and allow int and string which
have a clear equality comparison. Now allowing anything which is not float
assuming it will work later on in v.reclass where types generated by db_sqltype_to_Ctype()
are used. (These types are the simplified C-like equivalents of SQL types
which distinguish only str, int, double, and datetime.) Consequently, datetime columns
will fail in v.reclass and not in this condition which is suboptimal, but allows more
column types to work with v.dissolve.