wxGUI: Remove calls of unicode function which don't work in Python 3
This removes all direct calls of unicode() function which is not
present in Python 3.

The str() function is used instead to convert from exceptions/errors to strings.
The column iteration works without any conversion in Python 3.

Additionaly, this renames variables e to error and removes a todo note to
fix exceptional case in a better way (not clear enough to be useful).