wxGUI/dbmgr: Show proper error when table parsing failed (#1112)
When parsing of a row fails, instead of trying to show error message
(and getting the GUI in a partially frozen state), raise exception
which the caller code expects and shows the message in a window
and allows user to continue working (although no table content is displayed).

The message also contains more details and suggests most likely cause (and fix).

Explicitly return None to make the intention clear.

This is a quick-fix. The real fix would be to use real CSV or JSON format and
parser as noted in the comment.