v.db.select: Handle all formats through option, add CSV
Instead of separate flags for JSON and vertical output format, use format option
so -j is replaced by format=json, etc. The formats are no longer spread among
the other flags and a written command is more readble. It is also more explicit
since the default format is now represented in the interface and vertical output
clearly stands out as yet another alternative.

The default format is named plain, alternative names are be text, ascii and default.

The user of an option does not change anything major in the code in terms of output
except for use of ints/booleans instead of directly accessing the flags' answers.
The parameter handling changed significatly. The standard exclude function
cannot be used and custom checks and messages are needed.

This adds another format, CSV, as format=csv, i.e., without need to add another
flag (e.g., -c and -v were already taken). Most of the code was already there
as basic structure was the original plain default output and quoting was already
done for the JSON output. The original format was not a proper CSV because it
lacked handling quoting of fields which for common CSV parsers takes care of new line
characters and separators in values/cells. All non-numeric columns are quoted
(with two quotes representing a quote in the text).
The quotes used are double quotes, the separator is drived by the horizontal/field
separator option which is unfortunate because it is not a comma by default.
CSV output is available even for the -r flag (minimal region/bbox).

The JSON output of NULL values is fixed, writing JSON literal null.