lib: Increase max size of SQL statements
Increases length of SQL statements in db.execute and elsewhere (DB_SQL_MAX)
from 1024 * pow(2, 3) to 1024 * pow(2, 6). Longer SQL statements are needed
for long attribute values, but also for tables with a lot of columns,
esp. when this table needs to be dropped and re-created during v.db.dropcolumn
(which is the original motivation for this change).

The new value is much higher than the current one and the SQL in the
v.db.dropcolumn case above. If every occurence of DB_SQL_MAX in the
code (77) would be allocated at the same time, it would be still
under 5 MiB.