Change default color for rendering vectors (gray->blue) [news]
Use light and dark blue as defaults in d.vect and in the default GUI settings
for rendering vector maps.

The d.vect values are applied in d.mon and with direct rendering,
the GUI values when user does not have user settings yet and
when user settings is saved for the first time (since then it is driven
by value there, so this needs settings to not exists to take its effect).