wxGUI: After creating new location, switch to PERMANENT and display infobar for first-time user (#1183)
After creating a new Location from Data tab, GUI switches to its PERMANENT mapset.

If the previous location was the demolocation (i.e., it was the active location when user created the new location), an info bar appears offering to import data using Import vector data or Import raster data buttons. This is assumed to be a second info bar which appears to the user navigating through the GUI from the the demolocation (world location).

This follows 37b39db0b967f513b8f636f1efb134655fe1aa19 (#1078).

The demolocation detection is based on the location name.

A signal is used in order to trigger the info bar for any location creation done in the Data tab (button, first-time user intro info bar, context menu).