wxGUI: New, simplified, unified window titles (#1216)
The 'lmgr' window is considered the main one, so it is no longer
called (GRASS GIS) Layer Manager, but just GRASS GIS. This fits well with this
window being the main GRASS GIS component for many users, although it is not
technically completely precise.

The lmgr window title previously included only workspace name after its base title.
Now it includes workspace name (if applicable), followed by
mapset/location (previously in map display title) and that is followed by the base title,
separated by dashes. This follows common practice of document before application name,
or from most specific information to most general. This also brings the titles closer
to what a single window interface would use. It also may improve the situation reported by a beginner:
*Once I open the GRASS GIS console...it opens another application called Layer Manager*, reported in
[Trac Ticket 3474](https://trac.osgeo.org/grass/ticket/3474).

The version number of is no longer included in the title. This is again a common practice
and it makes making release screenshots easier. Bug reports need a more complete info anyway.
Users which have multiple installations are likely advanced anyway and can always check if needed.

Map display does not include mapset/location at the end as before,
nor GRASS GIS as prefix. Titles of tabs in the Display tab are fully synchronized
with the titles regardless of the title being the default or custom,
so the tabs now say Map Display, not just Display. This simplifies the code
significantly and clarifies the interface. (Small change to rename item
in menu is included for further clarification.)

Database manager which includes vector name in the title follows the same practice
as the main window lmgr, i.e., document before tool/app name.
The read-only info is included with more common spelling and standard case
(applied in the table tab title).

Dialogs started from main GUI don't show GRASS GIS as first part of the title.
This basic title is usually set as default in the frame.
The corresponding g.gui.* tools (if any) set their own title using
a tool - program name syntax and are considered to be somewhat standalone,
so they include GRASS GIS after a dash.

The GUI forms/dialogs for modules no longer include keywords in (square) brackets.
Module name is therefore more highlighted as the actual tool name.
Keywords are not meant to be used this way. (A useful title might be menu item
label, but that's not readily available.)

wxgui.py now sets the application name and vendor which improves how the program
is titled on some platforms. (Making no distinction between internal and display
name because there is no current use case for that.)

Use Histogram Tool instead of Histogramming Tool because Histogram tool is much more
common (74500 Google results) than Histogramming Tool (567). It is also shorter and easier to spell.
Leaving the word histogramming in statusbar where it is used as a verb.

The d.histogram based tool title now has the module name in square brackets
(preceded by two spaces) following the practice from menus and Modules tab.