wxGUI/mapdisp: Always respect user-set background color in Map Display (#1167)
The background color from user settings was applied only when actual rendering
happened. So with no layers at the start and when all unchecked the color was
the hardcoded white.

Now the color is always the same although it comes from different sources
(backround by default and g.pnmcomp result when displaying data).

This creates consistent behavior and it is a plus for the dark theme support
because when user manually the dark background to fit with the dark theme,
the big empty white area is no longer displayed when no data is there.

This applies to all components which use mapwin.BufferedMapWindow,
e.g., Map Swipe.