db.univar: Ignore rows with NULLs with -e
When value in a row is NULL, basic statistics ignore the rows
(e.g., not counted in N), however,
prior to this PR, the extended stats (-e) were processing all rows.
This was causing two issues:
1. db.univar/v.db.univar failed with 'could not convert string to float', or
2. wrong results when number of NULLs happeded to be less than N/4.

Now the NULL values (represented as empty strings) are always skipped in
the same was as for the basic stats,
i.e., not being counted towards the quartiles anymore (which is what was
causing the wrong results when number of NULL values was low).
With where conditions such as 'IS NOT NULL' neither of these issues emerged.
Now a similar condition is in the code for extended stasts too.

This PR refactors the rstrip calls, so they are in the code only once per loop.

The test now tests a run with both basic and extended stats with data
which contains NULL values. Actual values are not tested as before,
but setup and teardown steps are cleaned up (e.g. not trying to delete
elevation raster).