Move grass package content to directory named grass
Before this change, the code of the Python package called grass which comes
with GRASS GIS is hidden among C libraries in lib/python.
This may look like it is a C library to use or to communicate with Python,
not a Python package.

For Python packages, the standard naming is a directory which matches the name of the package.
This allows, for example, to import files directly from the source code.

This PR moves content of lib/python to python/grass so that the name of the directory
matches the name of the resulting package (currently placed in etc/python).

This does not change Sphinx which continues to ignore the top-level grass package
(this is future work).

Fixes #1100.