Flake8 fixes for utils and man
This applies fixes for basic Flake8 issues to Python scripts in utils and man.
Number of issues is ignored in the config file; ideally to be resolved later.

This adds a Flake8 config file to the top level directory and runs Flake8 from there
ignoring directories checked separatelly and files not yet fixed.
However, it checks any other Python file which shows up in the source code outside
of the standard directories (e.g., a script for creating visuals for a module doc).