parser: Accept #-space-% to allow writing PEP8-compliant Python code
Accept '# %' in parser script header definition because '#%' is against PEP8
(each line of a block comment starts with a # and a single space).
Only minimal changes to achive this support are included.

'#%' is still supported fully supported and remains in documentation.
The full switch to '# %' is left for v8.

Partial backport of 635224115c5ee9c29a9356210c3b03e5511ef6dc (#1287).