Remove perlcritic silencing.
Use statements are moved to line 22. Previous silencing of perlcritic on line 76 is now removed. It was intitally added as when implementing best practices for the open statement on line 76 i was getting
a error for an unopened filehandle. This was due to '<F>' needing to be '<>'. It is fixed now and perlcritic is not silenced in the file.