r.geomorphon: remove multires-specific code
This code does not work now. As the original developer explains, there
is no point in fixing it, because the intended use case can be
implemented with scripts around r.geomorphon instead of C code in it.

That said, leave multires.c be, it has been unused for a long time, but
the function in it might belong somewhere else.

This change addresses bug #1355.