WMS: replace broken URLs with alternative WMS
- both NASA GIBS and irs.gis-lab.info are no longer reachable
- replaced with global WMS by terrestris and Sentinel-2 by EOX

TODO: it is not clear to me if the EOX tile service is supported at all.

Alternative suggestions are welcome, ideally WMS services with global coverage.

Note that the list shown in the GUI may be extended by the local list in `$HOME/.grass8/wxWS` if present (i.e., old WMS records to be removed there)