v.perturb: Require either explicit seed or generated seed
This moves the modules to the standard followed by r.mapcalc where we require
user to either provide seed or to ask for a seed with -s flag
(-s results in a different seed and thus different results everytime).

The original code used fixed seed by default and required explicit
change to get a different 'random' result.
The GRASS standard G_srand48_auto() function is used which returns the seed
which is in turn used in the custom random number generator code in the module.