libpython: Make parallel region test somewhat more robust
Test introduced in #638 uses raster maps to determine if the subprocess was
successfully executed. However, in GitHub Actions this fails randomly
with a random number of rasters missing.

This PR makes the test wait for couple seconds before counting the rasters
in case the file system needs to catch up. If there is an actual error,
waiting won't fix it, so the test will still do its job.

If running and rerunning the tests here is successful, it suggests
that it is a good fix at least in the sense of accomodating
the given environment.

Alternative fix for the test would be to replace Bash by Python
which would allow leaving out the raster creation check (which is
replacing return code check here) and provide more control.

On one hand, we want to make the test run, on the other,
we want to have the actual code robust, but maybe writing 50
rasters in parallel in a small VM in GitHub Actions is not
a use case we need to cover in code and thus modifying the test is enough.