libgis: Distinguish dirs and objects in dir creation
The current G_make_mapset_element() is used in several places to create
nested directories such as grid3/some_raster3d_name for particular maps/objects.
However, it seems to me that that's very different from what G_make_mapset_element()
should be doing which is creating directories (such as fcell, vector) for these
objects.

This PR introduces new API functions which make distinction between these scenarios.
One function is for what most of the original API calls element that is the directory
for objects of the same type. The other is for the objects themselves.

The three new functions replace two existing ones.
Two replace G_make_mapset_element() covering the two cases and one replaces
G_make_mapset_element_tmp() covering only the common directory case
simply because I didn't see the tmp version to be used in that other way.

Distinguishing the two scenarios allows the code to handle differently a race
condition during creation when another process creates the same directory
between the failed access call and mkdir call.

In case of directory such as fcell or vector, all we asked
for was that the directory exists at the end of the function call which
will be fulfilled even when mkdir happened in another process.

Same race condition for directories such as vector/vector_map_name
is actuall problem and reporting it sooner rather than later is advantageous in
detecting the issue (that is two processes are trying to create map of the same name).

I think part of the issue is the unclear naming. Element sometimes refers to
the particular file such as fcell/raster_map_name, but here it refers to the
directory of these files such as fcell. This new API does not use great names,
but the idea is to refer to element group directories (such as cell) and specific
element files or directories (raster map piece, sub-directory for a specific vector map).