pygrass: Add update parameters method to Module
This adds a new update method which can update the parameters (and generally attributes) of the Module.
The current practice is to use the function call operator to do the update. This seems to go against
the function call operator semantics as the module is not executed unless explicit run_=True is passed.

This PR is not changing the current behavior, but it adds a better way of expressing the same idea.

The name update is based on update function in Python dict since this seems to be a similar case.