Add list of commits for git blame to ignore (#1391)
This adds a list of commits which one may want to ignore when doing git blame
(git blame file.c --ignore-revs-file .git-blame-ignore-revs).

Formatting changes of Python (with Black), the initial C formatting commit,
and couple of other commits are included.
Includes Black formatting changes up to 21.5b1 update.

The intention is an ignore list for investigating 'standard' code. It won't work when
an error was introduced by the formatting change (e.g., by fixing some additional lines
manually) or when dealing with CI checks changed in the same commit.

Includes basic documentation of what the file is for.