rst/interpfl lib: change devi arg to bool (#1271)
Output deviation argument to IL_init_params_2d() and in the
interp_params struct, was defined as FILE*. In practice a char* of the
file path was given to the function in 'v.surf.rst/main.c' and this
pointer's only use was to check whether it was NULL or not. It was used
as a flag. This changes the argument to be a bool.

Addresses a -Wincompatible-pointer-types compiler warning.