ci: Print version info for Ubuntu and CentOS runs (#1832)
To clearly show what version was installed and what dependencies are used, this prints version info at the end of Ubuntu and CentOS workflows (regardless of build status so possibly failing when GRASS GIS does not build correctly). This includes dependencies and extended Python version info using the Python set in the GRASS_PYTHON variable.